requirement.txt
strawberry-graphql==0.78.0
uvicorn==0.15.0
beautifulsoup4==4.9.3
requests==2.25.1
nltk==3.5
pandas==1.2.0
sklearn==0.0